Enforce -Wall -Wextra -Werror in tests

See merge request redox-os/relibc!434

+++ tests/ctype.c
@@ -301,7 +301,7 @@ size_t num_test_cases = sizeof(test_cases) / sizeof(struct test_case);
 int main(void) {
     int retval = EXIT_SUCCESS;
 
-    for(int i = 0; i < num_test_cases; ++i) {
+    for(size_t i = 0; i < num_test_cases; ++i) {
         struct test_case tc = test_cases[i];
         CHECK_TEST(tc, isalnum, retval);
         CHECK_TEST(tc, isalpha, retval);
