+++ tests/wctype/towupper.c
@@ -1,11 +1,12 @@
 #include <stdio.h>
+#include <stddef.h>
 #include <wchar.h>
 #include <wctype.h>
 
 int main() {
     wchar_t *str = L"HaLf WiDe ChAr StRiNg!\n";
     fputws(str, stdout);
-    for (int i = 0; i < wcslen(str); i++) {
+    for (size_t i = 0; i < wcslen(str); i++) {
         putwchar(towctrans(str[i], wctrans("toupper")));
     }
     return 0;
